+++ frontend_vue/src/components/base/BaseMessageBox.vue
